refactor: extract algorithm CLI module
